Fix shader program

+++ src/main/java/com/tttsaurus/saurus3d/common/core/shader/ShaderProgram.java
@@ -172,7 +172,7 @@ public class ShaderProgram implements Comparable<ShaderProgram>
         else if (TypeUtils.isIntOrWrappedInt(clazz))
             v = (float)((int)value);
         else if (TypeUtils.isDoubleOrWrappedDouble(clazz))
-            v = (float)value;
+            v = (float)((double)value);
         return v;
     }
     private static double getDouble(Class<?> clazz, Object value)
